# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AiSOC Platform API
  description: |
    Open-source AI Security Operations Center (v5.2.0).

    Autonomous threat detection, investigation, and response — with UEBA behavioral
    analytics, honeytoken deception, purple-team adversary emulation, enterprise RBAC,
    immutable audit logging, the v5.2 Investigation Ledger (signed agent actions),
    Ambient Copilot, Responder PWA, public evaluation harness, MCP server, and
    SOC 2 / ISO 27001 / NIST CSF compliance dashboards.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained from `POST /api/v1/auth/login`.
    Service-to-service calls may use an API key via `X-API-Key` header.

    ## Multi-tenancy
    All data is isolated via PostgreSQL Row-Level Security. The `tenant_id` is derived
    from the authenticated token and cannot be overridden by the caller.

    ## Rate limiting
    Default: 1000 req/min per API key; 100 req/min per user token.
    Limits are surfaced via `X-RateLimit-*` response headers.
  version: 5.2.0
  contact:
    name: AiSOC Community
    url: https://github.com/beenuar/AiSOC/security/advisories/new
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
paths:
...
  /api/v1/api-keys:
    post:
      tags:
      - api-keys
      summary: Create Api Key
      description: 'Create a new scoped API key for the current tenant.


        The raw key value is returned **only once** in this response.'
      operationId: create_api_key_api_v1_api_keys_post
      parameters:
      - name: current_user
        in: query
        required: true
        schema:
          title: Current User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApiKeyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - api-keys
      summary: List Api Keys
      description: List all API keys for the current tenant (raw keys are never returned).
      operationId: list_api_keys_api_v1_api_keys_get
      security:
      - HTTPBearer: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKeyOut'
                title: Response List Api Keys Api V1 Api Keys Get
  /api/v1/api-keys/{key_id}:
    get:
      tags:
      - api-keys
      summary: Get Api Key
      description: Get metadata for a specific API key.
      operationId: get_api_key_api_v1_api_keys__key_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Key Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - api-keys
      summary: Update Api Key
      description: Update an API key's name, scopes, or expiry.
      operationId: update_api_key_api_v1_api_keys__key_id__patch
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Key Id
      - name: current_user
        in: query
        required: true
        schema:
          title: Current User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateApiKeyRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - api-keys
      summary: Revoke Api Key
      description: Revoke (deactivate) an API key. The key can no longer be used for
        authentication.
      operationId: revoke_api_key_api_v1_api_keys__key_id__delete
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Key Id
      - name: current_user
        in: query
        required: true
        schema:
          title: Current User
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiKeyOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        prefix:
          type: string
          title: Prefix
        scopes:
          items:
            type: string
          type: array
          title: Scopes
        is_active:
          type: boolean
          title: Is Active
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        last_used_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Used At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - prefix
      - scopes
      - is_active
      - expires_at
      - last_used_at
      - created_at
      title: ApiKeyOut
...
    CreateApiKeyRequest:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
          description: Human-readable name
        scopes:
          items:
            type: string
          type: array
          title: Scopes
          description: Permission scopes granted to this key
        expires_in_days:
          anyOf:
          - type: integer
            maximum: 3650.0
            minimum: 1.0
          - type: 'null'
          title: Expires In Days
          description: TTL in days; None = never expires
      type: object
      required:
      - name
      title: CreateApiKeyRequest
    CreateApiKeyResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        key:
          type: string
          title: Key
          description: Raw API key — shown ONCE, store securely
        prefix:
          type: string
          title: Prefix
        scopes:
          items:
            type: string
          type: array
          title: Scopes
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - key
      - prefix
      - scopes
      - expires_at
      - created_at
      title: CreateApiKeyResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UpdateApiKeyRequest:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Name
        scopes:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Scopes
        expires_in_days:
          anyOf:
          - type: integer
            maximum: 3650.0
            minimum: 1.0
          - type: 'null'
          title: Expires In Days
          description: Reset TTL from now; 0 = never
      type: object
      title: UpdateApiKeyRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer